# openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.2
  title: Simple store
  license:
    name: MIT
servers:
  - url: http://localhost/api/v1
paths:
  /products:
    get:
      summary: Search products (paginated)
      operationId: searchProducts
      tags:
        - products
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            default: 10
            type: integer
            format: int32
        - name: page
          in: query
          description: WHic page we want to obtain
          required: false
          schema:
            type: integer
            format: int32
        - name: category
          in: query
          description: Filter products by category
          required: false
          schema:
            type: string
        - name: name
          in: query
          description: Filter product by name (LIKE 'name%')
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A paged array of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCollectionOfProducts"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SearchLimitError:
                  value:
                    code : 525
                    message : "Search limit cannot exceed 100 items."
    post:
      summary: Create a product
      operationId: createProduct
      tags:
        - products
      requestBody:
        description: Product information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Null response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{productCode}:
    get:
      summary: Info for a specific product
      operationId: getProductByCode
      tags:
        - products
      parameters:
        - name: productCode
          in: path
          required: true
          description: The code of the product to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update product by code
      operationId: updateProductByCode
      tags:
        - products
      parameters:
        - name: productCode
          in: path
          required: true
          description: The code of the product to retrieve
          schema:
            type: string
      requestBody:
        description: Product information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Null response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete product by code
      operationId: deleteProductByCode
      tags:
        - products
      parameters:
        - name: productCode
          in: path
          required: true
          description: The code of the product to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Null response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # cart paths
  /cart:
    get:
      summary: Get cart items
      operationId: getCart
      tags:
        - cart
      responses:
        '200':
          description: A list of cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: update cart content
      operationId: updateCart
      tags:
        - cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cart'
      responses:
        '200':
          description: Null response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Cart:
      type: array
      items:
        $ref:  "#/components/schemas/CartItem"
    CartItem:
      type: object
      required:
        - code
        - amount
      properties:
        code:
          type: string
        amount:
          type: integer
          format: int32
    Product:
      type: object
      required:
        - code
        - name
        - category
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        productImage:
          # relative path to product image
          type: string
          #later will be changed into ProductImage[]
        price:
          type: number
          format: float
        created_at:
          type: string
          format: date
        updated_at:
          nullable: true
          type: string
          format: date
    PaginatedCollectionOfProducts:
      type: object
      properties:
        items:
          $ref:  "#/components/schemas/ProductCollection"
        pagination:
          $ref:  "#/components/schemas/Pagination"
    ProductCollection:
      type: array
      items:
        $ref: "#/components/schemas/Product"
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          format: int32
          default: 10
          minimum: 1
          maximum: 100
        page:
          type: integer
          format: int32
          minimum: 1
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string